+++ src/result.rs
@@ -339,7 +339,7 @@ impl EvalAltResult {
     /// Consume the current `EvalAltResult` and return a new one with the specified `Position`
     /// if the current position is `Position::None`.
     #[inline(always)]
-    pub(crate) fn new_position(mut self: Box<Self>, new_position: Position) -> Box<Self> {
+    pub(crate) fn fill_position(mut self: Box<Self>, new_position: Position) -> Box<Self> {
         if self.position().is_none() {
             self.set_position(new_position);
         }